Split SettingsManager into per-domain mixins (Phase 4a: god-file decomposition)
Problem: graphlink_settings_store.py was a 1626-line god-file - one
SettingsManager class owning ~10 genuinely distinct settings domains
(persistence/migration engine, general app prefs, Ollama, Llama.cpp, cloud
API providers, MCP servers, plugin grants, recipes, harness trusted dirs,
pricing overrides) with no internal boundaries. Imported by ~45 files across
the backend and test suite, always as exactly one of two names
(SettingsManager, KEEP_EXISTING_SECRET) - a wide blast radius that made this
the highest-risk item to get precisely right.

Change: split into a new settings_store/ package (10 files, one per domain,
1852 lines total) mirroring backend/domain/graph.py's own established
mixin-composition pattern for SceneDocument (BranchOps/GroupOps/LayoutOps/
CommandOps). graphlink_settings_store.py itself stays a real top-level file
at its unchanged import path - now a 97-line composition root holding just
the KEEP_EXISTING_SECRET sentinel, the class-level constants every mixin
shares, and:

  class SettingsManager(
      PersistenceOps, GeneralSettingsOps, OllamaSettingsOps, LlamaCppSettingsOps,
      CloudProviderSettingsOps, McpSettingsOps, PluginGrantsOps, RecipesOps,
      HarnessSettingsOps, PricingSettingsOps,
  )

Every one of the 112 original methods was relocated, verified AST-identical
to its original body (111/112 byte-for-byte; the 1 exception,
set_api_settings, needed one added deferred import to avoid a circular
import back to KEEP_EXISTING_SECRET's home module - confirmed the only
diff). No public API changed; no caller needed to change.

Also fixed a real packaging gap the split introduced: pyproject.toml's
[tool.setuptools.packages.find] include list only allowlisted "backend*"
and "graphlink_plugins*" - the new settings_store package would have been
silently excluded from the built wheel (confirmed via find_packages()
before/after, and by actually building the wheel and installing it into a
clean venv both before the fix, where the settings_store/*.py files were
absent from the wheel, and after, where they're present and the full
backend.agents -> graphlink_settings_store -> settings_store.* import
chain works from the installed package).

Test plan:
- Full backend suite from the repo root (python -m pytest -q): 3073 passed,
  19 skipped, 1 failed - the one failure is the same pre-existing,
  environment-sensitive timing flake in test_providers.py already tracked
  as unrelated to recent work (a tight 2.0s assertion this machine's
  current load doesn't always meet; reproduces with every change in this
  branch reverted).
- ruff check . and mypy: both clean.
- A real construct/save/get/set/migrate/KEEP_EXISTING_SECRET round trip
  against a temp session.dat, run directly against the new code.
- python -m build --wheel followed by installing the built wheel into a
  clean venv and importing backend.agents/graphlink_settings_store/
  settings_store.persistence from it, matching CI's build-check job.
