+++ settings_store/__init__.py
@@ -1,0 +1,1 @@
+"""The settings-store package: SettingsManager's per-domain mixins."""
